+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - backport to C++11 since Sailfish OS (as of 3.0.3)  does not support C++17 yet
 
+### Fixed
+- display widget in check not being hidden when a new data item is being checked
+
 ## [1.0.0] - 2019-05-25
 First public release!